requirements.txt
argon2-cffi==20.1.0
asgiref @ file:///tmp/build/80754af9/asgiref_1594338739818/work
bcrypt==3.1.7
certifi==2020.6.20
cffi==1.14.1
Django==3.0.3
django-crispy-forms==1.9.2
djangorestframework==3.11.1
Faker @ file:///tmp/build/80754af9/faker_1593440881412/work
numpy==1.19.1
pandas==1.1.1
Pillow==7.2.0
pycparser==2.20
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
